feat: se reestructuró la configuración de Alembic y se añadieron nuevos archivos
- Se actualizó la ubicación del script de Alembic en `alembic.ini`.
- Se eliminaron archivos obsoletos de Alembic y se crearon nuevos archivos de configuración.
- Se implementó un nuevo archivo `env.py` con soporte para múltiples proveedores de base de datos.
- Se añadió un archivo `README.md` con instrucciones para el uso de Alembic.
