+++ alembic.ini
@@ -1,5 +1,5 @@
 [alembic]
-script_location = alembic
+script_location = app/core/db/alembic
 sqlalchemy.url = sqlite:///macti.db
 
 [loggers]
