--- alembic/env.py
@@ -1,86 +1,0 @@
-import sys
-from logging.config import fileConfig
-from pathlib import Path
-
-from sqlalchemy import engine_from_config, pool
-
-from alembic import context
-
-# Añadir la raíz del proyecto al path para poder importar `app`
-PROJECT_ROOT = str(Path(__file__).resolve().parents[1])
-if PROJECT_ROOT not in sys.path:
-    sys.path.insert(0, PROJECT_ROOT)
-
-# this is the Alembic Config object, which provides
-# access to the values within the .ini file in use.
-config = context.config
-
-# Interpret the config file for Python logging if present.
-if config.config_file_name:
-    fileConfig(config.config_file_name)
-
-# Import la configuración y la metadata de la aplicación
-import app.shared.models  # noqa: F401,E402  # importar modelos para registrarlos
-from app.core.db.database import SQLALCHEMY_DATABASE_URL, Base  # noqa: E402
-
-target_metadata = Base.metadata
-
-
-def run_migrations_offline():
-    """Run migrations in "offline" mode.
-
-    This configures the context with just a URL
-    and not an Engine, though an Engine is acceptable
-    here as well.  By skipping the Engine creation
-    we don't even need a DBAPI to be available.
-
-    Calls to context.execute() here emit the given string to the
-    script output.
-    """
-
-    url = SQLALCHEMY_DATABASE_URL
-    context.configure(
-        url=url,
-        target_metadata=target_metadata,
-        literal_binds=True,
-        dialect_opts={"paramstyle": "named"},
-    )
-
-    with context.begin_transaction():
-        context.run_migrations()
-
-
-def run_migrations_online():
-    """Run migrations in "online" mode.
-
-    In this scenario we need to create an Engine
-    and associate a connection with the context.
-    """
-
-    # Ensure Alembic knows the SQLAlchemy URL
-    config.set_main_option("sqlalchemy.url", SQLALCHEMY_DATABASE_URL)
-    # get_section may return None; ensure we pass a dict to engine_from_config
-    configuration = config.get_section(config.config_ini_section) or {}
-
-    connectable = engine_from_config(
-        configuration,
-        prefix="sqlalchemy.",
-        poolclass=pool.NullPool,
-    )
-
-    with connectable.connect() as connection:
-        # enable compare_type so Alembic detects column type changes
-        context.configure(
-            connection=connection,
-            target_metadata=target_metadata,
-            compare_type=True,
-        )
-
-        with context.begin_transaction():
-            context.run_migrations()
-
-
-if context.is_offline_mode():
-    run_migrations_offline()
-else:
-    run_migrations_online()
