--- alembic/versions/3e3b4458d6ae_initial.py
@@ -1,74 +1,0 @@
-"""initial
-
-Revision ID: 3e3b4458d6ae
-Revises: None
-Create Date: 2026-05-20 21:39:51.963549
-"""
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "3e3b4458d6ae"
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    conn = op.get_bind()
-    # Comprobar y eliminar índice/tabla solo si existen (SQLite)
-    idx_exists = conn.execute(
-        sa.text(
-            "SELECT name FROM sqlite_master WHERE type='index' AND name='ix_MCT_user_courses_id'"
-        )
-    ).fetchone()
-    if idx_exists:
-        op.drop_index(op.f("ix_MCT_user_courses_id"), table_name="MCT_user_courses")
-
-    tbl_exists = conn.execute(
-        sa.text(
-            "SELECT name FROM sqlite_master WHERE type='table' AND name='MCT_user_courses'"
-        )
-    ).fetchone()
-    if tbl_exists:
-        op.drop_table("MCT_user_courses")
-    # SQLite doesn't support ALTER TYPE; use batch_alter_table for compatibility
-    with op.batch_alter_table("MCT_verification_tokens") as batch_op:
-        batch_op.alter_column(
-            "token",
-            existing_type=sa.NUMERIC(),
-            type_=sa.String(length=36),
-            existing_nullable=False,
-        )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("MCT_verification_tokens") as batch_op:
-        batch_op.alter_column(
-            "token",
-            existing_type=sa.String(length=36),
-            type_=sa.NUMERIC(),
-            existing_nullable=False,
-        )
-    op.create_table(
-        "MCT_user_courses",
-        sa.Column("id", sa.INTEGER(), nullable=False),
-        sa.Column("auth_id", sa.INTEGER(), nullable=False),
-        sa.Column("course_full_name", sa.VARCHAR(), nullable=True),
-        sa.Column("groups", sa.VARCHAR(), nullable=True),
-        sa.Column("status", sa.VARCHAR(length=8), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["auth_id"],
-            ["MCT_auth.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_MCT_user_courses_id"), "MCT_user_courses", ["id"], unique=False
-    )
-    # ### end Alembic commands ###
