--- alembic/versions/bcee208a2b56_initial.py
@@ -1,54 +1,0 @@
-"""initial
-
-Revision ID: bcee208a2b56
-Revises: 3e3b4458d6ae
-Create Date: 2026-05-20 21:46:05.039429
-"""
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "bcee208a2b56"
-down_revision = "3e3b4458d6ae"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - adjusted for SQLite ###
-    conn = op.get_bind()
-    tbl_exists = conn.execute(
-        sa.text(
-            "SELECT name FROM sqlite_master WHERE type='table' AND name='MCT_verification_tokens'"
-        )
-    ).fetchone()
-    if tbl_exists:
-        with op.batch_alter_table("MCT_verification_tokens") as batch_op:
-            batch_op.alter_column(
-                "token",
-                existing_type=sa.VARCHAR(length=36),
-                type_=sa.String(length=36),
-                existing_nullable=False,
-            )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - adjusted for SQLite ###
-    conn = op.get_bind()
-    tbl_exists = conn.execute(
-        sa.text(
-            "SELECT name FROM sqlite_master WHERE type='table' AND name='MCT_verification_tokens'"
-        )
-    ).fetchone()
-    if tbl_exists:
-        with op.batch_alter_table("MCT_verification_tokens") as batch_op:
-            batch_op.alter_column(
-                "token",
-                existing_type=sa.String(length=36),
-                type_=sa.VARCHAR(length=36),
-                existing_nullable=False,
-            )
-    # ### end Alembic commands ###
