+++ app/core/db/alembic/versions/bf21c0e49cbd_initial.py
@@ -1,0 +1,27 @@
+"""initial
+
+Revision ID: bf21c0e49cbd
+Revises: None
+Create Date: 2026-06-18 17:15:06.030783
+"""
+
+import sqlalchemy as sa  # noqa: F401
+from alembic import op  # noqa: F401
+
+# revision identifiers, used by Alembic.
+revision = "bf21c0e49cbd"
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
